# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Journey AI API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

export type JourneyStatus = (typeof JourneyStatus)[keyof typeof JourneyStatus];

export const JourneyStatus = {
  active: "active",
  paused: "paused",
  completed: "completed",
  archived: "archived",
} as const;

export interface Task {
  id: string;
  phaseId: string;
  orderIndex: number;
  title: string;
  description: string;
  isCompleted: boolean;
  /** @nullable */
  completedAt: string | null;
}

export interface Phase {
  id: string;
  journeyId: string;
  orderIndex: number;
  title: string;
  description: string;
  estimatedDuration: string;
  tasks: Task[];
}

export interface Journey {
  id: string;
  goalText: string;
  title: string;
  summary: string;
  status: JourneyStatus;
  createdAt: string;
  updatedAt: string;
  phases: Phase[];
}

export interface JourneySummary {
  id: string;
  title: string;
  summary: string;
  status: JourneyStatus;
  createdAt: string;
  totalTasks: number;
  completedTasks: number;
  progressPercent: number;
  /** @nullable */
  nextTaskTitle: string | null;
}

export interface PhaseProgress {
  phaseId: string;
  title: string;
  totalTasks: number;
  completedTasks: number;
  progressPercent: number;
}

export interface JourneyProgress {
  journeyId: string;
  totalTasks: number;
  completedTasks: number;
  progressPercent: number;
  /** @nullable */
  currentPhaseId: string | null;
  /** @nullable */
  currentPhaseTitle: string | null;
  phaseBreakdown: PhaseProgress[];
}

export type ActivityEventType =
  (typeof ActivityEventType)[keyof typeof ActivityEventType];

export const ActivityEventType = {
  task_completed: "task_completed",
  task_uncompleted: "task_uncompleted",
  task_added: "task_added",
  coach_reply: "coach_reply",
} as const;

export interface ActivityEvent {
  id: string;
  type: ActivityEventType;
  title: string;
  timestamp: string;
}

export interface DashboardOverview {
  totalJourneys: number;
  activeJourneys: number;
  completedJourneys: number;
  totalTasks: number;
  completedTasks: number;
  currentStreak: number;
}

/**
 * @nullable
 */
export type CreateJourneyBodyExperienceLevel =
  | (typeof CreateJourneyBodyExperienceLevel)[keyof typeof CreateJourneyBodyExperienceLevel]
  | null;

export const CreateJourneyBodyExperienceLevel = {
  beginner: "beginner",
  intermediate: "intermediate",
  advanced: "advanced",
} as const;

export interface CreateJourneyBody {
  /** @minLength 4 */
  goalText: string;
  /**
   * Optional desired timeframe (e.g. "3 months", "1 year")
   * @nullable
   */
  timeline?: string | null;
  /** @nullable */
  experienceLevel?: CreateJourneyBodyExperienceLevel;
}

export interface UpdateJourneyBody {
  title?: string;
  status?: JourneyStatus;
}

export interface CreateTaskBody {
  /** @minLength 1 */
  title: string;
  description?: string;
}

export interface UpdateTaskBody {
  title?: string;
  description?: string;
  isCompleted?: boolean;
}

export type CoachMessageRole =
  (typeof CoachMessageRole)[keyof typeof CoachMessageRole];

export const CoachMessageRole = {
  user: "user",
  coach: "coach",
} as const;

export interface CoachMessage {
  id: string;
  journeyId: string;
  role: CoachMessageRole;
  content: string;
  createdAt: string;
}

export interface SendCoachMessageBody {
  /** @minLength 1 */
  content: string;
}

export interface CoachReply {
  userMessage: CoachMessage;
  coachMessage: CoachMessage;
}