# lib/api-zod/src/generated/types/coachMessage.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Journey AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CoachMessageRole } from "./coachMessageRole";

export interface CoachMessage {
  id: string;
  journeyId: string;
  role: CoachMessageRole;
  content: string;
  createdAt: Date;
}