# lib/api-zod/src/generated/types/coachReply.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Journey AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CoachMessage } from "./coachMessage";

export interface CoachReply {
  userMessage: CoachMessage;
  coachMessage: CoachMessage;
}