# lib/api-zod/src/generated/types/createJourneyBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Journey AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateJourneyBodyExperienceLevel } from "./createJourneyBodyExperienceLevel";

export interface CreateJourneyBody {
  /** @minLength 4 */
  goalText: string;
  /**
   * Optional desired timeframe (e.g. "3 months", "1 year")
   * @nullable
   */
  timeline?: string | null;
  /** @nullable */
  experienceLevel?: CreateJourneyBodyExperienceLevel;
}