# lib/api-zod/src/generated/types/journeyProgress.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Journey AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PhaseProgress } from "./phaseProgress";

export interface JourneyProgress {
  journeyId: string;
  totalTasks: number;
  completedTasks: number;
  progressPercent: number;
  /** @nullable */
  currentPhaseId: string | null;
  /** @nullable */
  currentPhaseTitle: string | null;
  phaseBreakdown: PhaseProgress[];
}