# lib/api-zod/src/generated/types/phaseProgress.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Journey AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PhaseProgress {
  phaseId: string;
  title: string;
  totalTasks: number;
  completedTasks: number;
  progressPercent: number;
}