# lib/api-zod/src/generated/types/sendCoachMessageBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Journey AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SendCoachMessageBody {
  /** @minLength 1 */
  content: string;
}