# lib/api-zod/src/generated/types/task.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Journey AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Task {
  id: string;
  phaseId: string;
  orderIndex: number;
  title: string;
  description: string;
  isCompleted: boolean;
  /** @nullable */
  completedAt: Date | null;
}