# lib/api-zod/src/generated/types/updateJourneyBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Journey AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { JourneyStatus } from "./journeyStatus";

export interface UpdateJourneyBody {
  title?: string;
  status?: JourneyStatus;
}